chore(2025-10): add naive solution for part 2 (too slow)
Add a naive BFS graph search solution for part 2. It is way too slow.

The solution needs to be fast enough to explore graphs of depth around
50.

I added a TODO comments for the optimizations that can be done to
improve the performance using A* algorithm (estimate distance to
goal using a metric), rather than pure BFS that walks all nodes in all
directions.